requirements.txt
# ===========================
# FastAPI & Web Framework
# ===========================
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
numpy==1.26.4
# ===========================
# PDF & Image Processing
# ===========================
PyMuPDF==1.23.8
Pillow==10.1.0
opencv-python==4.8.1.78

# ===========================
# AI
# ===========================
google-generativeai==0.3.1

# ===========================
# Database
# ===========================
sqlalchemy==2.0.23

# Choose ONE database driver:
# PostgreSQL (Production)
psycopg2-binary==2.9.9
# ===========================
 prometheus-client==0.17.1
# ===========================
# Cache
# ===========================
redis==5.0.1

# ===========================
# Utilities
# ===========================
python-dotenv==1.0.0
pydantic-settings==2.1.0

# ===========================
# HTTP Client
# ===========================
httpx==0.25.2